internal: Fix Scala 3 related warnings
