+++ airframe-benchmark/src/main/scala/wvlet/airframe/benchmark/http/ServiceSyncClient.scala
@@ -13,26 +13,22 @@
  */
 package wvlet.airframe.benchmark.http
 
-/**
-  */
 import wvlet.airframe.http._
 import wvlet.airframe.http.client.SyncClient
 
-import scala.collection.immutable.Map
-
-class ServiceSyncClient[Req, Resp](private val client: HttpSyncClient[Req, Resp]) extends AutoCloseable {
-  override def close(): Unit               = { client.close() }
-  def getClient: HttpSyncClient[Req, Resp] = client
-  object Greeter {
-    def hello(name: String, requestFilter: Req => Req = identity): String = {
-      client.postOps[Map[String, Any], String](
-        resourcePath = s"/wvlet.airframe.benchmark.http.Greeter/hello",
-        Map("name" -> name),
-        requestFilter = requestFilter
-      )
-    }
-  }
-}
+//class ServiceSyncClient[Req, Resp](private val client: HttpSyncClient[Req, Resp]) extends AutoCloseable {
+//  override def close(): Unit               = { client.close() }
+//  def getClient: HttpSyncClient[Req, Resp] = client
+//  object Greeter {
+//    def hello(name: String, requestFilter: Req => Req = identity): String = {
+//      client.postOps[Map[String, Any], String](
+//        resourcePath = s"/wvlet.airframe.benchmark.http.Greeter/hello",
+//        Map("name" -> name),
+//        requestFilter = requestFilter
+//      )
+//    }
+//  }
+//}
 
 class NewServiceSyncClient(private val client: SyncClient) extends AutoCloseable {
   override def close(): Unit = { client.close() }
