+++ airframe-control/.jvm/src/main/scala/wvlet/airframe/control/Shell.scala
@@ -66,7 +66,7 @@ object Shell extends LogSupport {
     // retrieve child processes
     val pb = prepareProcessBuilder(s"ps -o pid -p $pid | sed 1d", inheritIO = true)
     // TODO: Use lazyLines when Scala 2.12 dropped
-    for (line <- Process(pb).lineStream) {
+    for (line <- Process(pb).!!.linesIterator) {
       val childPID = line.trim.toInt
       killTree(childPID, signal)
     }
