+++ airframe-sql/src/main/scala/wvlet/airframe/sql/catalog/DataType.scala
@@ -156,17 +156,17 @@ object DataType extends LogSupport {
   case object AnyType     extends PrimitiveType("any")
   case object NullType    extends PrimitiveType("null")
 
-  case object BooleanType                                   extends PrimitiveType("boolean")
-  abstract class NumericType(override val typeName: String) extends PrimitiveType(typeName)
-  case object ByteType                                      extends NumericType("byte")
-  case object ShortType                                     extends NumericType("short")
-  case object IntegerType                                   extends NumericType("integer")
-  case object LongType                                      extends NumericType("long")
+  case object BooleanType                      extends PrimitiveType("boolean")
+  abstract class NumericType(typeName: String) extends PrimitiveType(typeName)
+  case object ByteType                         extends NumericType("byte")
+  case object ShortType                        extends NumericType("short")
+  case object IntegerType                      extends NumericType("integer")
+  case object LongType                         extends NumericType("long")
 
-  abstract class FractionType(override val typeName: String) extends NumericType(typeName)
-  case object FloatType                                      extends FractionType("float")
-  case object RealType                                       extends FractionType("real")
-  case object DoubleType                                     extends FractionType("double")
+  abstract class FractionType(typeName: String) extends NumericType(typeName)
+  case object FloatType                         extends FractionType("float")
+  case object RealType                          extends FractionType("real")
+  case object DoubleType                        extends FractionType("double")
 
   case class CharType(length: Option[DataType])    extends DataType("char", length.toSeq)
   case object StringType                           extends PrimitiveType("string")
